Refactor load neo4j Objects to include previous generate_model. Some code movedf to new itinary_gen
