+++ requirements.txt
@@ -1,6 +1,6 @@
-fastapi==0.100.0
-pydantic==2.0.3
-uvicorn
+fastapi==0.100.1
+pydantic==2.1.1
+uvicorn==0.23.2
 psycopg2-binary
 python-dotenv~=1.0.0
 pytest==7.4.0
@@ -12,7 +12,7 @@ gitpython~=3.1.18
 dpath==2.1.6
 geopy==2.3.0
 pandas==2.0.3
-numpy==1.25.1
+numpy==1.25.2
 matplotlib==3.7.2
 seaborn==0.12.2
 contextily==1.3.0
