+++ source/api/requirements.txt
@@ -1,7 +1,7 @@
-fastapi==0.100.0
-starlette==0.27.0
+fastapi==0.100.1
+starlette==0.31.0
 httpx
-pydantic==2.0.3
+pydantic==2.1.1
 python-dotenv
 uvicorn==0.23.1
 psycopg2-binary
